修正AppServer存放文件目录
