+++ internal/k8s/service/uesr/app_service.go
@@ -1,26 +1,60 @@
-/*
- * MIT License
- *
- * Copyright (c) 2024 Bamboo
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- *
- */
-
 package uesr
+
+import (
+	"context"
+	"fmt"
+	"github.com/GoSimplicity/AI-CloudOps/internal/k8s/client"
+	"github.com/GoSimplicity/AI-CloudOps/internal/k8s/dao/admin"
+	admin2 "github.com/GoSimplicity/AI-CloudOps/internal/k8s/service/admin"
+	"github.com/GoSimplicity/AI-CloudOps/internal/model"
+	"go.uber.org/zap"
+)
+
+type AppService interface {
+	CreateInstanceOne(ctx context.Context, instance *model.K8sInstanceRequest) error
+}
+type appService struct {
+	dao               admin.ClusterDAO
+	client            client.K8sClient
+	l                 *zap.Logger
+	deploymentService admin2.DeploymentService
+	svcService        admin2.SvcService
+}
+
+func NewAppService(dao admin.ClusterDAO, client client.K8sClient, l *zap.Logger, deploymentService admin2.DeploymentService, svcService admin2.SvcService) AppService {
+	return &appService{
+		dao:               dao,
+		client:            client,
+		l:                 l,
+		deploymentService: deploymentService,
+		svcService:        svcService,
+	}
+}
+
+func (a *appService) CreateInstanceOne(ctx context.Context, instance *model.K8sInstanceRequest) error {
+	// 1.创建deployment请求参数
+	deploymentRequest := &model.K8sDeploymentRequest{
+		ClusterId:       instance.ClusterId,
+		Namespace:       instance.Namespace,
+		DeploymentNames: instance.DeploymentNames,
+		DeploymentYaml:  instance.DeploymentYaml,
+	}
+	// 调用deploymentService的CreateDeployment方法创建deployment
+	err := a.deploymentService.CreateDeployment(ctx, deploymentRequest)
+	if err != nil {
+		return fmt.Errorf("failed to create Deployment: %w", err)
+	}
+	// 2.创建service请求参数
+	svcRequest := &model.K8sServiceRequest{
+		ClusterId:    instance.ClusterId,
+		Namespace:    instance.Namespace,
+		ServiceNames: instance.ServiceNames,
+		ServiceYaml:  instance.ServiceYaml,
+	}
+	// 调用svcService的CreateService方法创建service
+	err = a.svcService.CreateService(ctx, svcRequest)
+	if err != nil {
+		return fmt.Errorf("failed to create Service: %w", err)
+	}
+	return nil
+}
