+++ pkg/di/wire_gen.go
@@ -13,6 +13,7 @@ import (
 	"github.com/GoSimplicity/AI-CloudOps/internal/k8s/client"
 	"github.com/GoSimplicity/AI-CloudOps/internal/k8s/dao/admin"
 	admin2 "github.com/GoSimplicity/AI-CloudOps/internal/k8s/service/admin"
+	user2  "github.com/GoSimplicity/AI-CloudOps/internal/k8s/service/uesr"
 	api4 "github.com/GoSimplicity/AI-CloudOps/internal/not_auth/api"
 	service4 "github.com/GoSimplicity/AI-CloudOps/internal/not_auth/service"
 	api6 "github.com/GoSimplicity/AI-CloudOps/internal/prometheus/api"
@@ -111,7 +112,7 @@ func InitWebServer() *Cmd {
 	k8sYamlTaskHandler := api5.NewK8sYamlTaskHandler(logger, yamlTaskService)
 	yamlTemplateService := admin2.NewYamlTemplateService(yamlTemplateDAO, yamlTaskDAO, k8sClient, logger)
 	k8sYamlTemplateHandler := api5.NewK8sYamlTemplateHandler(logger, yamlTemplateService)
-	k8sAppService := admin2.NewAppService(clusterDAO, k8sClient, logger,deploymentService,svcService)
+	k8sAppService := user2.NewAppService(clusterDAO, k8sClient, logger,deploymentService,svcService)
 	k8sAppHandler := api5.NewK8sAppHandler(logger, k8sAppService)
 	alertManagerEventDAO := alert.NewAlertManagerEventDAO(db, logger, userDAO)
 	scrapePoolDAO := scrape.NewScrapePoolDAO(db, logger, userDAO)
